+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Version 4.25.0-SNAPSHOT
 -------------
 
+ADDED:
+ 
+- created new method for get an Integer from properties object with a given properties key
+
 CHANGED:
 
 - javadoc extended and improved